Move dataset and metadata write ops into mutations tree
