+++ src/adapter/graphql/src/mutations/mod.rs
@@ -7,7 +7,15 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod auth;
-mod tasks;
-pub(crate) use auth::*;
-pub(crate) use tasks::*;
+mod auth_mut;
+mod dataset_metadata_mut;
+mod dataset_mut;
+mod datasets_mut;
+mod metadata_chain_mut;
+mod tasks_mut;
+pub(crate) use auth_mut::*;
+pub(crate) use dataset_metadata_mut::*;
+pub(crate) use dataset_mut::*;
+pub(crate) use datasets_mut::*;
+pub(crate) use metadata_chain_mut::*;
+pub(crate) use tasks_mut::*;
