+++ src/adapter/graphql/src/queries/datasets/datasets.rs
@@ -105,202 +105,6 @@ impl Datasets {
         let account = Account::mock();
         self.by_account_impl(ctx, account, page, per_page).await
     }
-
-    /// Creates a new empty dataset
-    async fn create_empty(
-        &self,
-        ctx: &Context<'_>,
-        account_id: AccountID,
-        dataset_kind: DatasetKind,
-        dataset_name: DatasetName,
-    ) -> Result<CreateDatasetResult> {
-        match self
-            .create_from_snapshot_impl(
-                ctx,
-                account_id,
-                odf::DatasetSnapshot {
-                    name: dataset_name.into(),
-                    kind: dataset_kind.into(),
-                    metadata: Vec::new(),
-                },
-            )
-            .await?
-        {
-            CreateDatasetFromSnapshotResult::Success(s) => Ok(CreateDatasetResult::Success(s)),
-            CreateDatasetFromSnapshotResult::NameCollision(e) => {
-                Ok(CreateDatasetResult::NameCollision(e))
-            }
-            CreateDatasetFromSnapshotResult::InvalidSnapshot(_)
-            | CreateDatasetFromSnapshotResult::Malformed(_)
-            | CreateDatasetFromSnapshotResult::UnsupportedVersion(_)
-            | CreateDatasetFromSnapshotResult::MissingInputs(_) => unreachable!(),
-        }
-    }
-
-    /// Creates a new dataset from provided DatasetSnapshot manifest
-    async fn create_from_snapshot(
-        &self,
-        ctx: &Context<'_>,
-        account_id: AccountID,
-        snapshot: String,
-        snapshot_format: MetadataManifestFormat,
-    ) -> Result<CreateDatasetFromSnapshotResult> {
-        use odf::serde::DatasetSnapshotDeserializer;
-
-        let snapshot = match snapshot_format {
-            MetadataManifestFormat::Yaml => {
-                let de = odf::serde::yaml::YamlDatasetSnapshotDeserializer;
-                match de.read_manifest(snapshot.as_bytes()) {
-                    Ok(snapshot) => snapshot,
-                    Err(e @ odf::serde::Error::SerdeError { .. }) => {
-                        return Ok(CreateDatasetFromSnapshotResult::Malformed(
-                            MetadataManifestMalformed {
-                                message: e.to_string(),
-                            },
-                        ));
-                    }
-                    Err(odf::serde::Error::UnsupportedVersion(e)) => {
-                        return Ok(CreateDatasetFromSnapshotResult::UnsupportedVersion(
-                            e.into(),
-                        ))
-                    }
-                    Err(e @ odf::serde::Error::IoError { .. }) => return Err(e.int_err().into()),
-                }
-            }
-        };
-
-        self.create_from_snapshot_impl(ctx, account_id, snapshot)
-            .await
-    }
-
-    // TODO: Multi-tenancy
-    // TODO: Multi-tenant resolution for derivative dataset inputs (should it only
-    // work by ID?)
-    #[allow(unused_variables)]
-    #[graphql(skip)]
-    async fn create_from_snapshot_impl(
-        &self,
-        ctx: &Context<'_>,
-        account_id: AccountID,
-        snapshot: odf::DatasetSnapshot,
-    ) -> Result<CreateDatasetFromSnapshotResult> {
-        let dataset_repo = from_catalog::<dyn domain::DatasetRepository>(ctx).unwrap();
-
-        let result = match dataset_repo
-            .create_dataset_from_snapshot(None, snapshot)
-            .await
-        {
-            Ok(result) => {
-                let dataset = Dataset::from_ref(ctx, &result.dataset_handle.as_local_ref()).await?;
-                CreateDatasetFromSnapshotResult::Success(CreateDatasetResultSuccess { dataset })
-            }
-            Err(domain::CreateDatasetFromSnapshotError::NameCollision(e)) => {
-                CreateDatasetFromSnapshotResult::NameCollision(CreateDatasetResultNameCollision {
-                    account_name: e.alias.account_name.map(|a| a.into()),
-                    dataset_name: e.alias.dataset_name.into(),
-                })
-            }
-            Err(domain::CreateDatasetFromSnapshotError::InvalidSnapshot(e)) => {
-                CreateDatasetFromSnapshotResult::InvalidSnapshot(
-                    CreateDatasetResultInvalidSnapshot { message: e.reason },
-                )
-            }
-            Err(domain::CreateDatasetFromSnapshotError::MissingInputs(e)) => {
-                CreateDatasetFromSnapshotResult::MissingInputs(CreateDatasetResultMissingInputs {
-                    missing_inputs: e
-                        .missing_inputs
-                        .into_iter()
-                        .map(|r| r.to_string())
-                        .collect(),
-                })
-            }
-            Err(domain::CreateDatasetFromSnapshotError::Internal(e)) => return Err(e.into()),
-        };
-
-        Ok(result)
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-// CreateDatasetResult
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(Interface, Debug, Clone)]
-#[graphql(field(name = "message", ty = "String"))]
-pub enum CreateDatasetResult {
-    Success(CreateDatasetResultSuccess),
-    NameCollision(CreateDatasetResultNameCollision),
-}
-
-#[derive(Interface, Debug, Clone)]
-#[graphql(field(name = "message", ty = "String"))]
-pub enum CreateDatasetFromSnapshotResult {
-    Success(CreateDatasetResultSuccess),
-    NameCollision(CreateDatasetResultNameCollision),
-    Malformed(MetadataManifestMalformed),
-    UnsupportedVersion(MetadataManifestUnsupportedVersion),
-    InvalidSnapshot(CreateDatasetResultInvalidSnapshot),
-    // TODO: This error should probably be generalized along with other
-    // errors that can occur during the metadata evolution
-    MissingInputs(CreateDatasetResultMissingInputs),
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(SimpleObject, Debug, Clone)]
-#[graphql(complex)]
-pub struct CreateDatasetResultSuccess {
-    pub dataset: Dataset,
-}
-
-#[ComplexObject]
-impl CreateDatasetResultSuccess {
-    async fn message(&self) -> String {
-        format!("Success")
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(SimpleObject, Debug, Clone)]
-#[graphql(complex)]
-pub struct CreateDatasetResultNameCollision {
-    pub account_name: Option<AccountName>,
-    pub dataset_name: DatasetName,
-}
-
-#[ComplexObject]
-impl CreateDatasetResultNameCollision {
-    async fn message(&self) -> String {
-        format!("Dataset with name '{}' already exists", self.dataset_name)
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(SimpleObject, Debug, Clone)]
-pub struct CreateDatasetResultInvalidSnapshot {
-    pub message: String,
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-#[derive(SimpleObject, Debug, Clone)]
-#[graphql(complex)]
-pub struct CreateDatasetResultMissingInputs {
-    // TODO: Input can be referenced by ID or by name
-    // do we need DatasetRef-style scalar for GQL?
-    pub missing_inputs: Vec<String>,
-}
-
-#[ComplexObject]
-impl CreateDatasetResultMissingInputs {
-    async fn message(&self) -> String {
-        format!(
-            "Dataset is referencing non-existing inputs: {}",
-            self.missing_inputs.join(", ")
-        )
-    }
 }
 
 ///////////////////////////////////////////////////////////////////////////////
