+++ src/adapter/graphql/src/queries/datasets/metadata_chain.rs
@@ -142,92 +142,6 @@ impl MetadataChain {
             total_count,
         ))
     }
-
-    /// Commits new event to the metadata chain
-    #[tracing::instrument(level = "info", skip_all)]
-    async fn commit_event(
-        &self,
-        ctx: &Context<'_>,
-        event: String,
-        event_format: MetadataManifestFormat,
-    ) -> Result<CommitResult> {
-        let event = match event_format {
-            MetadataManifestFormat::Yaml => {
-                let de = odf::serde::yaml::YamlMetadataEventDeserializer;
-                match de.read_manifest(event.as_bytes()) {
-                    Ok(event) => event,
-                    Err(e @ odf::serde::Error::SerdeError { .. }) => {
-                        return Ok(CommitResult::Malformed(MetadataManifestMalformed {
-                            message: e.to_string(),
-                        }))
-                    }
-                    Err(odf::serde::Error::UnsupportedVersion(e)) => {
-                        return Ok(CommitResult::UnsupportedVersion(e.into()))
-                    }
-                    Err(e @ odf::serde::Error::IoError { .. }) => return Err(e.int_err().into()),
-                }
-            }
-        };
-
-        let dataset = self.get_dataset(ctx).await?;
-
-        let result = match dataset
-            .commit_event(event, domain::CommitOpts::default())
-            .await
-        {
-            Ok(result) => CommitResult::Success(CommitResultSuccess {
-                old_head: result.old_head.map(Into::into),
-                new_head: result.new_head.into(),
-            }),
-            Err(domain::CommitError::ObjectNotFound(e)) => {
-                CommitResult::AppendError(CommitResultAppendError {
-                    message: format!("Event is referencing a non-existent object {}", e.hash),
-                })
-            }
-            Err(domain::CommitError::MetadataAppendError(e)) => {
-                CommitResult::AppendError(CommitResultAppendError {
-                    message: e.to_string(),
-                })
-            }
-            Err(e @ domain::CommitError::Internal(_)) => return Err(e.int_err().into()),
-        };
-
-        Ok(result)
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
-
-#[derive(Interface, Debug, Clone)]
-#[graphql(field(name = "message", ty = "String"))]
-pub enum CommitResult {
-    Success(CommitResultSuccess),
-    Malformed(MetadataManifestMalformed),
-    UnsupportedVersion(MetadataManifestUnsupportedVersion),
-    AppendError(CommitResultAppendError),
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
-
-#[derive(SimpleObject, Debug, Clone)]
-#[graphql(complex)]
-pub struct CommitResultSuccess {
-    pub old_head: Option<Multihash>,
-    pub new_head: Multihash,
-}
-
-#[ComplexObject]
-impl CommitResultSuccess {
-    async fn message(&self) -> String {
-        format!("Success")
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
-
-#[derive(SimpleObject, Debug, Clone)]
-pub struct CommitResultAppendError {
-    pub message: String,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////
